`ConnectivityViewController` parent `UIViewController`
Handle `deinit` initializing `PostManager` referencing `self`!
