+++ Sources/Views/PostView/PostView.swift
@@ -44,6 +44,12 @@ open class PostView: UIView, Poster {
     
     private func setup() {
         clipsToBounds = true
+        
+        // Make sure `postManger` is instantiated. This is encase `deinit` is
+        // is the first instance to reference `postManager`, lazily
+        // instantiating it, referencing `self`, which is being
+        // de-initialized...
+        let _ = self.postManager.isSerialQueue
     }
     
     // MARK: - IntrinsicContentSize
